@@ -5,7 +5,7 @@ git_source(:github) do |repo_name|
   "https://github.com/#{repo_name}.git"
 end
 
-ruby '2.4.1'
+ruby '2.6.6'
 
 gem 'bootstrap-sass', '~> 3.4.1'
 gem 'bootstrap3-datetimepicker-rails', '~> 4.17.47'
@@ -23,7 +23,7 @@ gem 'omniauth-github'
 gem 'pg', '~> 0.18'
 gem 'postmark-rails'
 gem 'puma', '~> 3.12'
-gem 'rails', '~> 5.2.2'
+gem 'rails', '~> 5.2.5'
 gem 'react-rails', '~> 2.2.1'
 gem 'sass-rails', '~> 5.0'
 gem 'uglifier', '>= 1.3.0'
@@ -36,7 +36,7 @@ end
 gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
 
 group :development, :test do
-  gem 'factory_girl_rails', '~> 4.0'
+  gem 'factory_bot_rails', '~> 4.0'
   gem 'pry-rails'
 end
 
