@@ -1,11 +1,13 @@
-FactoryGirl.define do
+FactoryBot.define do
 
   factory :user do
-    name 'John Galt'
-    email 'somewhere@colorado.com'
-    sequence :uid { |i| '90210' + i.to_s }
-    provider User::AUTH_PROVIDER_GITHUB
-    token 'averysecrettoken'
+    name { 'John Galt' }
+    email { 'somewhere@colorado.com' }
+    sequence :uid do |i|
+      '90210' + i.to_s
+    end
+    provider { User::AUTH_PROVIDER_GITHUB }
+    token { 'averysecrettoken' }
   end
 
   trait :reviewer do
